app/requirements.txt
arrow==0.13.1
binaryornot==0.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
Flask==1.0.2
Flask-Bootstrap4==4.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jinja2-time==0.2.0
MarkupSafe==1.1.1
pefile==2019.4.18
poyo==0.4.2
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.2
whichcraft==0.5.2
